reference/correspondence unification
